@@ -14,8 +14,9 @@
     "format:check": "prettier --check \"**/*.{ts,tsx,mdx}\" --cache"
   },
   "dependencies": {
-    "@cometh/connect-sdk": "^1.2.7",
-    "@cometh/marketplace-sdk": "2.1.0-develop",
+    "@cometh/connect-sdk": "^1.2.13",
+    "@cometh/connect-sdk-viem": "^1.1.7",
+    "@cometh/marketplace-sdk": "2.1.0",
     "@ethersproject/abi": "^5.7.0",
     "@ethersproject/bignumber": "^5.7.0",
     "@ethersproject/providers": "^5.7.2",
@@ -37,61 +38,61 @@
     "@radix-ui/react-toast": "^1.1.5",
     "@radix-ui/react-tooltip": "^1.0.7",
     "@rainbow-me/rainbowkit": "^2.0.1",
-    "@tanstack/react-query": "^5.17.12",
-    "@tanstack/react-query-devtools": "^5.17.12",
-    "@tanstack/react-table": "^8.11.6",
+    "@tanstack/react-query": "^5.25.0",
+    "@tanstack/react-query-devtools": "^5.25.0",
+    "@tanstack/react-table": "^8.13.2",
     "@traderxyz/nft-swap-sdk": "^0.33.0",
     "@web3-onboard/core": "^2.21.2",
     "@web3-onboard/injected-wallets": "^2.10.11",
-    "axios": "^1.6.5",
+    "axios": "^1.6.7",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
-    "cmdk": "^0.2.0",
+    "cmdk": "^0.2.1",
     "dayjs": "^1.11.10",
     "ethers": "^5.7.2",
-    "jotai": "^2.6.2",
+    "jotai": "^2.7.0",
     "lodash": "^4.17.21",
     "lucide-react": "^0.309.0",
     "luxon": "^3.4.4",
     "next": "^13.5.5",
     "next-themes": "^0.2.1",
-    "qs": "^6.11.2",
+    "qs": "^6.12.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-hook-form": "^7.49.3",
-    "react-intersection-observer": "^9.5.3",
+    "react-hook-form": "^7.51.0",
+    "react-intersection-observer": "^9.8.1",
     "react-spring": "^9.7.3",
     "sharp": "^0.33.2",
-    "tailwind-merge": "^2.2.0",
+    "tailwind-merge": "^2.2.1",
     "tailwindcss-animate": "^1.0.7",
-    "usehooks-ts": "^2.9.4",
-    "viem": "^2.1.1",
-    "wagmi": "^2.2.1",
+    "usehooks-ts": "^2.16.0",
+    "viem": "^2.7.20",
+    "wagmi": "^2.5.7",
     "zod": "^3.22.4"
   },
   "devDependencies": {
     "@ianvs/prettier-plugin-sort-imports": "^4.1.1",
-    "@next/bundle-analyzer": "^14.0.4",
+    "@next/bundle-analyzer": "^14.1.3",
     "@simbathesailor/use-what-changed": "^2.0.0",
     "@types/axios": "^0.14.0",
     "@types/lodash": "^4.14.202",
-    "@types/luxon": "^3.4.1",
-    "@types/node": "^20.11.4",
-    "@types/qs": "^6.9.11",
-    "@types/react": "^18.2.48",
-    "@types/react-dom": "^18.2.18",
+    "@types/luxon": "^3.4.2",
+    "@types/node": "^20.11.25",
+    "@types/qs": "^6.9.12",
+    "@types/react": "^18.2.64",
+    "@types/react-dom": "^18.2.21",
     "@typescript-eslint/parser": "^6.19.0",
-    "autoprefixer": "^10.4.16",
+    "autoprefixer": "^10.4.18",
     "encoding": "^0.1.13",
-    "eslint": "^8.56.0",
+    "eslint": "^8.57.0",
     "eslint-config-next": "13.0.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-tailwindcss": "^3.14.0",
-    "postcss": "^8.4.33",
-    "prettier": "^3.2.2",
-    "prettier-plugin-tailwindcss": "^0.5.11",
+    "eslint-plugin-react": "^7.34.0",
+    "eslint-plugin-tailwindcss": "^3.14.3",
+    "postcss": "^8.4.35",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.5.12",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5.3.3"
+    "typescript": "^5.4.2"
   }
 }
