@@ -1,8 +1,13 @@
 "use client"
 
-import { useConnectModal } from "@rainbow-me/rainbowkit"
+import { useCallback } from "react"
+import { manifest } from "@/manifests"
+import { marketplaceChain } from "@/providers/authentication/marketplaceWagmiChain"
+import { ComethConnectConnector } from "@cometh/connect-sdk-viem"
+import { ConnectButton, useConnectModal } from "@rainbow-me/rainbowkit"
 import { cx } from "class-variance-authority"
 import { WalletIcon } from "lucide-react"
+import { useConnect } from "wagmi"
 
 import { env } from "@/config/env"
 import { Button } from "@/components/ui/button"
@@ -12,7 +17,6 @@ import {
   DropdownMenuContent,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu"
-import { ConnectButton } from '@rainbow-me/rainbowkit';
 
 import { AccountWallet } from "./account-wallet"
 
@@ -30,6 +34,16 @@ export function SigninDropdown({
   isLinkVariant,
 }: SigninDropdownProps) {
   const { openConnectModal } = useConnectModal()
+  const { connect } = useConnect()
+
+  const handleComethConnectLogin = useCallback(async () => {
+    // const connector = new ComethConnectConnector({
+    //   chains: [marketplaceChain],
+    //   options: { apiKey: manifest.walletConnectProjectId },
+    // })
+    // await connect({ connector })
+  }, [connect])
+
   const wallets = [
     ...(env.NEXT_PUBLIC_COMETH_CONNECT_API_KEY
       ? [
@@ -37,7 +51,7 @@ export function SigninDropdown({
             name: "Cometh Connect",
             icon: `${env.NEXT_PUBLIC_BASE_PATH}/cometh-connect.png`,
             isComethWallet: true,
-            handleConnect: async () => {},
+            handleConnect: handleComethConnectLogin,
           },
         ]
       : []),
@@ -46,7 +60,6 @@ export function SigninDropdown({
       icon: `${env.NEXT_PUBLIC_BASE_PATH}/metamask.svg`,
       isComethWallet: false,
       handleConnect: () => {
-        console.log("handleConnect", openConnectModal)
         openConnectModal && openConnectModal()
       },
     },
@@ -69,7 +82,6 @@ export function SigninDropdown({
       </DropdownMenuTrigger>
       <DropdownMenuContent align="end" asChild>
         <Card className="p-4" style={{ width: "324px" }}>
-          <ConnectButton />
           <CardHeader className="mb-3 p-0">
             <CardTitle className="text-xl">Login</CardTitle>
           </CardHeader>
@@ -84,6 +96,10 @@ export function SigninDropdown({
               />
             ))}
           </CardContent>
+          {/* ConnectButton MUST BE included in the DOM for the modal to appear, but can be hidden.*/}
+          <div className="hidden">
+            <ConnectButton />
+          </div>
         </Card>
       </DropdownMenuContent>
     </DropdownMenu>
